refactor: reorganize directory structure with src/ layout
- Create src/ directory for all source code
- Move geometry_tutor/ and api/ to src/
- Move run_api_server.py to scripts/ directory
- Update all import paths to use src. prefix
- Simplify __init__.py files to avoid eager imports
- Fix sys.path modifications for new structure
