+++ backend/scripts/run_api_server.py
@@ -9,11 +9,11 @@ import argparse
 from pathlib import Path
 
 # Add the project root to Python path
-project_root = Path(__file__).parent
+project_root = Path(__file__).parent.parent
 sys.path.insert(0, str(project_root))
 
-from api.main import run_server
-from geometry_tutor.llm_utils import setup_environment
+from src.api.main import run_server
+from src.geometry_tutor.llm_utils import setup_environment
 
 
 def main():
