+++ backend/src/geometry_tutor/__init__.py
@@ -5,8 +5,6 @@ This package implements an AI-powered Geometry Tutor system using LangGraph
 and LLM-native reasoning designed for Vietnamese high school students.
 """
 
-from .tutor import GeometryTutor
-from .core import GraphState, create_initial_state
-
 __version__ = "0.1.0"
+# Lazy imports to avoid dependency issues during development
 __all__ = ["GeometryTutor", "GraphState", "create_initial_state"]
