+++ backend/src/geometry_tutor/cli.py
